// generated/addon/angular/src/app/api-models/cieb-addon-me-config.model.ts
/* Copyright (c) 2021 . All Rights Reserved. */

export class CiebAddonMeConfig {
  seqCodeId: number;
  countryCode: string;
  codeType: string;
  code1: string;
  codeDesc: string;
  code2: string;
  insertDatetime: Date;
  insertUser: string;
  insertProcess: string;
  updateDatetime: Date;
  updateUser: string;
  updateProcess: string;
}